==> notebooks/2 - data_mining_the_players_info_summary_dataset.ipynb <==
import pandas as pd

player_info_summary_dataset = pd.read_csv("../datasets/players_info_summary_new.csv")

without_indian_players =  player_info_summary_dataset[player_info_summary_dataset["country"] != "India"]

without_indian_players.info()

played_international =  without_indian_players.dropna(subset=["played_international"])

played_international

played_international.info()

players_with_country_na = played_international[played_international["country"].isna()]
players_with_country_na

- All 3 above players are Indian Players. So They will be removed 

cleaned_played_international = played_international.dropna(subset=["country"])
cleaned_played_international.info()
cleaned_played_international


bowling_style_na =  cleaned_played_international[cleaned_played_international["bowling_style"].isna()]
bowling_style_na

- All the players with bowling style missing are batters.

cleaned_played_international.to_csv('../datasets/international_played_foreign_players.csv', index=False)
