==> notebooks/4 -  scrapping_ipl_matches.ipynb <==
import pandas as pd
from selenium import webdriver
from selenium.webdriver.common.by import By
from selenium.webdriver.common.keys import Keys
import time
from selenium.webdriver.chrome.service import Service
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.support.ui import Select
from bs4 import BeautifulSoup

players_dataset = pd.read_csv('../datasets/international_played_foreign_players.csv')
players_dataset.head()

player_search_page_url = "https://stats.espncricinfo.com/ci/engine/stats/index.html"

test_df = players_dataset[(players_dataset["player_id"] <= 708) & (players_dataset["player_id"] > 703)]
test_df.head()


driver = webdriver.Chrome(service=Service(ChromeDriverManager().install()))

all_player_dfs = []

for index,row in players_dataset.iterrows():
    player_name = row["player_name"]
    player_id = row["player_id"]
    driver.get(player_search_page_url)
    print(player_name)
    
    # Locate the search input field and type the player name
    search_box = driver.find_element(By.NAME, "search")
    search_box.send_keys(player_name)
    
    # Locate and click the submit button
    submit_button = driver.find_element(By.NAME, "submit_analysis")
    submit_button.click()
    
#     time.sleep(1)
    
    players_tab = driver.find_element(By.ID, "player")
    players_tab.click()
    
#     time.sleep(1)
    
    t20_link = driver.find_element(By.LINK_TEXT, "Twenty20 matches player")
    t20_link.click()

    # T20 Query Page
    advanced_filter = driver.find_element(By.LINK_TEXT, "Advanced filter")
    advanced_filter.click()
    
    
    try:
        trophy_dropdown = driver.find_element(By.NAME, "trophy")
        select = Select(trophy_dropdown)
        select.select_by_value("117")
        
        players_dataset.at[index, 'played_ipl'] = 'Yes'

        match_by_match_radio_button = driver.find_element(By.XPATH, '//input[@type="radio" and @name="view" and @value="match"]')
        match_by_match_radio_button.click()

        submit_button = driver.find_element(By.NAME, "submit_results")
        submit_button.click()

        # Getting table data

        soup = BeautifulSoup(driver.page_source, 'html.parser')
        tables = soup.find_all('table', {'class': 'engineTable'})
        for table in tables:
            caption = table.find('caption')
            if caption and caption.text.strip() == "Match by match list":
                headers = [th.text.strip() for th in table.find('thead').find_all('th')]
                rows = []
                for row in table.find('tbody').find_all('tr'):
                    cells = [td.text.strip() for td in row.find_all('td')]

                    # Extract the match link from the column with title="view the scorecard for this row"
                    td_link = row.find('td', {'style': 'white-space: nowrap;'})
                    match_link = None  # Default to None if no link is found

                    if td_link:
                        link = td_link.find('a', {'title': 'view the scorecard for this row'})
                        if link:
                            match_link = link.get('href')  # Extract the link (relative)

                    # Add the match_link to the row
                    cells.append(match_link)

                    rows.append(cells)

                # Add match_link as a new column
                headers.append('match_link')  # Add 'match_link' to headers
                df = pd.DataFrame(rows, columns=headers)
                df['player_id'] = player_id  # Add player_id column
                all_player_dfs.append(df)
    except Exception as e:
        players_dataset.at[index, 'played_ipl'] = 'No'

final_df = pd.concat(all_player_dfs, ignore_index=True)    
final_df.to_csv('../datasets/international_played_foreign_players_ipl_matches_details.csv',index=False)
print("successful")
